feat: adding new access flow for react native - fixes IOS-222
Adds RNAccess.mm/h containing Swift module and SDK methods.
Attaches that module to React-Native through NativeModules.
Removes usage of RCTViewManager, it is not needed anymore as we don't want to create a Swift UIView
but we'd rather call AccessIOS methods directly and let it handle the layout.

Misc: temporarly uses a local version of AccessIOS (~2.3.+) because we need to use
`createReactNativePaywall(_:)` method that isn't in production yet.